fix(webui): move V3PlayerViewState/V3PlayerLabeledValue into playerViewStateModel
The extraction imported these types from types/v3-player, but they were
never defined there -- they lived as local interfaces in
usePlaybackOrchestrator.ts. tsc --noEmit (part of npm run build) failed;
Vitest's transform didn't catch the invalid import. Moved both interfaces
to playerViewStateModel.ts (where the shape is now actually constructed)
and re-exported them from the orchestrator for existing consumers
(V3PlayerView.tsx/.test.tsx).
